+++ R/aaa_ranges.R
@@ -47,8 +47,8 @@ range_validate <- function(object, range, ukn_ok = TRUE) {
   else
     ""
   if (length(range) != 2)
-    stop(
-      "`range` must have an upper and lower bound. ", ukn_txt, call. = FALSE
+    rlang::abort(
+      paste0("`range` must have an upper and lower bound. ", ukn_txt)
     )
 
   is_unk <- is_unknown(range)
@@ -57,17 +57,16 @@ range_validate <- function(object, range, ukn_ok = TRUE) {
 
   if (!ukn_ok) {
     if (any(is_unk))
-      stop("Cannot validate ranges when they contains 1+ unknown values.",
-           call. = FALSE)
+      rlang::abort("Cannot validate ranges when they contains 1+ unknown values.")
     if (!any(is_num))
-      stop("`range` should be numeric.", call. = FALSE)
+      rlang::abort("`range` should be numeric.")
 
     # TODO check with transform
   } else {
     if (any(is_na[!is_unk]))
-      stop("Value ranges must be non-missing.", ukn_txt, call. = FALSE)
+      rlang::abort("Value ranges must be non-missing.", ukn_txt)
     if (any(!is_num[!is_unk]))
-      stop("Value ranges must be numeric.", ukn_txt, call. = FALSE)
+      rlang::abort("Value ranges must be numeric.", ukn_txt)
   }
   range
 }
@@ -86,7 +85,7 @@ range_get <- function(object, original = TRUE) {
 #' @rdname range_validate
 range_set <- function(object, range) {
   if (length(range) != 2)
-    stop("`range` should have two elements.")
+    rlang::abort("`range` should have two elements.")
   if (inherits(object, "quant_param")) {
     object <-
       new_quant_param(
@@ -99,7 +98,7 @@ range_set <- function(object, range) {
         label = object$label
       )
   } else {
-    stop("`object` should be a 'quant_param' object", call. = FALSE)
+    rlang::abort("`object` should be a 'quant_param' object")
   }
   object
 }