+++ R/param_svm.R
@@ -5,6 +5,7 @@
 #' @inheritParams Laplace
 #' @details
 #' These are used by `parsnip::svm_rbf()` and `parsnip::svm_poly()`
+#' @importFrom lifecycle deprecate_soft
 #' @examples
 #' cost()
 #' margin()
@@ -23,6 +24,19 @@ cost <- function(range = c(-10, -1), trans = log2_trans()) {
 #' @rdname cost
 #' @export
 margin <- function(range = c(0, .2), trans = NULL) {
+  deprecate_soft(
+    when = "0.0.5",
+    what = "margin()",
+    with = "svm_margin()",
+    details = "Details: https://github.com/tidymodels/dials/issues/85"
+  )
+
+  svm_margin(range = range, trans = trans)
+}
+
+#' @rdname cost
+#' @export
+svm_margin <- function(range = c(0, .2), trans = NULL) {
   new_quant_param(
     type = "double",
     range = range,