+++ tests/testthat/test_encode_unit.R
@@ -1,5 +1,4 @@
-library(testthat)
-library(dials)
+
 
 context("unit encoding")